+++ src/QPlus/core/Math.js
@@ -1,0 +1,9 @@
+Math.randomIntBetween = function (min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+Math.randomBetween = function (min, max) {
+  return Math.random() * (max - min) + min;
+};
