+++ src/QPlus/core/SimpleTilemap.js
@@ -1,0 +1,3 @@
+export default function SimpleTilemap() {
+  this.initialize.apply(this, arguments);
+}
