+++ src/QPlus/managers/SceneManager.js
@@ -1,0 +1,5 @@
+var Alias_SceneManager_updateManagers = SceneManager.updateManagers;
+SceneManager.updateManagers = function () {
+  Alias_SceneManager_updateManagers.call(this);
+  _QPlus.update();
+};
