+++ src/QPlus/scenes/Scene_Base.js
@@ -1,0 +1,14 @@
+var Alias_Scene_Base_initialize = Scene_Base.prototype.initialize;
+Scene_Base.prototype.initialize = function () {
+  Alias_Scene_Base_initialize.call(this);
+  this._waitListeners = [];
+  if (this.mixinWait()) {
+    _QPlus.mixinWait(this);
+  }
+};
+
+Scene_Base.prototype.mixinWait = function () {
+  // In your own scene, have this return true to be able to
+  // use the .wait(duration, callback) function
+  return false;
+};
